Changed waitForAbortableDelay.ts (line 1) and updated the three demo examples to use it, removing the duplicated local wait helpers
